+++ EP1/montaMatrizQuadrada.h
@@ -1,7 +1,7 @@
 #ifndef MONTAMATRIZQUADRADA_H_INCLUDED
 #define MONTAMATRIZQUADRADA_H_INCLUDED
 
-float** montaMatrizQuadrada(char fileName[100], int* tam);
+double** montaMatrizQuadrada(char fileName[100], int* tam);
 
 
 #endif // MONTAMATRIZQUADRADA_H_INCLUDED
