+++ EP1/montaVetor.h
@@ -1,7 +1,7 @@
 #ifndef MONTAMAVETOR_H_INCLUDED
 #define MONTAVETOR_H_INCLUDED
 
-float* montaVetor(char fileName[100], int* tam);
+double* montaVetor(char fileName[100], int* tam);
 
 
 #endif // MONTAVETOR_H_INCLUDED
